requirements.txt
backports.shutil-get-terminal-size==1.0.0
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
configparser==3.5.0
decorator==4.3.0
enum34==1.1.6
flake8==3.5.0
futures==3.2.0
google-api-core==1.1.2
google-auth==1.4.1
google-cloud-language==1.0.2
googleapis-common-protos==1.5.3
grpcio==1.12.0
idna==2.6
ipython==5.7.0
ipython-genutils==0.2.0
mccabe==0.6.1
pathlib2==2.3.2
pexpect==4.5.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
ptyprocess==0.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pytz==2018.4
requests==2.18.4
rsa==3.4.2
scandir==1.7
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
